atualizando

+++ src/br/edu/ifrn/schoolplan/classes/Disciplina.java
@@ -6,7 +6,7 @@ public class Disciplina {
     private String nome;
     private Color cor;
     private String professor;
-    private int n1, n2, n3, n4;
+    private double n1, n2, n3, n4;
     
     public Disciplina(String nome, String professor, Color cor) {
         this.nome = nome;
@@ -14,35 +14,35 @@ public class Disciplina {
         this.cor = cor;
     }
 
-    public int getN1() {
+    public double getN1() {
         return n1;
     }
 
-    public void setN1(int n1) {
+    public void setN1(double n1) {
         this.n1 = n1;
     }
 
-    public int getN2() {
+    public double getN2() {
         return n2;
     }
 
-    public void setN2(int n2) {
+    public void setN2(double n2) {
         this.n2 = n2;
     }
 
-    public int getN3() {
+    public double getN3() {
         return n3;
     }
 
-    public void setN3(int n3) {
+    public void setN3(double n3) {
         this.n3 = n3;
     }
 
-    public int getN4() {
+    public double getN4() {
         return n4;
     }
 
-    public void setN4(int n4) {
+    public void setN4(double n4) {
         this.n4 = n4;
     }
